# contracts/openapi/trainer-core.yaml
openapi: 3.1.0
info:
  title: MindTrain Training Core API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
paths:
  /api/v1/sessions:
    post:
      operationId: createTrainingSession
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                domainId: {type: string, default: java-backend}
                schedulerProvider:
                  type: string
                  enum: [weighted]
                  default: weighted
                  description: 稳定技术标识符；weighted 的中文展示名为“加权调度”。
      responses:
        '200': {description: Session created}
  /api/v1/sessions/{sessionId}/assignments/next:
    post:
      operationId: getNextAssignment
      parameters:
        - {name: sessionId, in: path, required: true, schema: {type: string}}
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Safe assignment or generation request with a Core-selected question type, difficulty, and knowledge point
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [assignment, generation_required, no_available_items, session_complete]
                  generationProfile:
                    oneOf:
                      - $ref: '#/components/schemas/GenerationProfile'
                      - type: 'null'
  /api/v1/assignments/{assignmentId}/attempts:
    post:
      operationId: submitChoiceAnswer
      parameters:
        - {name: assignmentId, in: path, required: true, schema: {type: string}}
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answer]
              properties:
                answer: {type: string}
      responses:
        '200': {description: Deterministic grading and explanation}
  /api/v1/assignments/{assignmentId}/reject:
    post:
      operationId: rejectGeneratedQuestion
      description: 仅允许拒绝尚未作答且属于当前用户 Session 的 AI 临时题；物理删除 Assignment、QuestionVersion 和 Question，并恢复主问题新题额度。
      parameters:
        - {name: assignmentId, in: path, required: true, schema: {type: string}}
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': {description: 临时题已物理删除，可以请求替代题}
        '404': {description: Assignment 不存在或不属于当前用户}
        '409': {description: 题目已经作答，或不是可拒绝的 AI 临时题}
  /api/v1/sessions/{sessionId}/interactions:
    post:
      operationId: recordInteraction
      parameters:
        - {name: sessionId, in: path, required: true, schema: {type: string}}
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': {description: Interaction appended without consuming the assignment}
  /api/v1/sessions/{sessionId}/finish:
    post:
      operationId: finishTrainingSession
      parameters:
        - {name: sessionId, in: path, required: true, schema: {type: string}}
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': {description: Session completed}
  /api/v1/candidates:
    post:
      operationId: createCandidateQuestion
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': {description: Session-scoped temporary candidate accepted; answering activates it for ordinary scheduling}
  /api/v1/questions/{questionId}/revisions:
    post:
      operationId: reviseActiveQuestion
      description: 基于当前可复习题创建不可变的新版本；旧 Assignment 和 Attempt 继续引用原版本。
      parameters:
        - {name: questionId, in: path, required: true, schema: {type: string}}
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersion, changes, reason]
              properties:
                expectedVersion: {type: integer, minimum: 1}
                changes:
                  type: object
                  description: 只提交需要变更的题目字段；不允许修改 id、version、status 和审计字段。
                reason: {type: string, minLength: 1}
                sourceAssignmentId: {type: string}
                model: {type: string}
                promptVersion: {type: string}
      responses:
        '200': {description: 新版本已保存并写入修订审计记录}
        '400': {description: 修订字段、来源 Assignment 或题目结构无效}
        '409': {description: 题目不是 active 可复习题，或 expectedVersion 已过期}
  /api/v1/reports/overview:
    get:
      operationId: getLearningReport
      description: 默认要求 Bearer Token；仅当实例显式开启公开看板时允许匿名读取聚合统计。
      responses:
        '200':
          description: 学习与内容概览；包含按 reportingTimeZone 统计的今日主问题数、每日目标和调度配额。
  /api/v1/schedulers/backlog:
    get:
      operationId: getSchedulerBacklog
      description: 默认要求 Bearer Token；仅当实例显式开启公开看板时允许匿名读取调度积压。
      responses:
        '200': {description: Due backlog and new-item allowance}
  /api/v1/settings/training:
    get:
      operationId: getTrainingSettings
      description: 读取数据库中的应用级训练配置；需要 Bearer Token。
      responses:
        '200':
          description: 当前训练配置，reviewBudget 等于 questionCount 减 newBudget。
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TrainingSettings'}
    put:
      operationId: updateTrainingSettings
      description: 更新应用级训练配置；新建 Session 随即使用新配置，已有 Session 保持创建时的 targetCount。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrainingSettings'
      responses:
        '200':
          description: 配置已保存
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TrainingSettings'}
        '400': {description: 配置值或时区无效}
  /api/v1/catalog/imports/preview:
    post:
      operationId: previewKnowledgeCatalogImport
      description: 校验并保存由用户所选资料整理出的训练领域与知识点草案，不修改正式领域、知识点、关系或来源资产。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CatalogPreviewRequest'}
      responses:
        '200':
          description: 草案、哈希、校验结果和差异
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CatalogImport'}
  /api/v1/catalog/imports/{importId}:
    get:
      operationId: getKnowledgeCatalogImport
      parameters:
        - {name: importId, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Catalog import preview}
        '404': {description: Import not found for the authenticated user}
  /api/v1/catalog/imports/{importId}/apply:
    post:
      operationId: applyKnowledgeCatalogImport
      description: 用户确认后，使用预览返回的精确 proposalHash 事务性写入正式 Catalog。
      parameters:
        - {name: importId, in: path, required: true, schema: {type: string}}
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [proposalHash]
              properties:
                proposalHash: {type: string, pattern: '^[0-9a-f]{64}$'}
      responses:
        '200': {description: Catalog applied or an idempotent prior result}
        '409': {description: Hash mismatch, conflict, invalid or already rejected import}
  /api/v1/catalog/imports/{importId}/reject:
    post:
      operationId: rejectKnowledgeCatalogImport
      parameters:
        - {name: importId, in: path, required: true, schema: {type: string}}
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': {description: Preview rejected without changing active catalog}
components:
  schemas:
    CatalogPreviewRequest:
      type: object
      required: [libraryId, proposal]
      properties:
        libraryId: {type: string}
        proposal:
          type: object
          required: [domains, topics, relations, sources]
          description: 只包含知识体系及来源元数据；禁止绝对路径、原文和抽取文本。
          properties:
            domains: {type: array, items: {type: object}}
            topics: {type: array, items: {type: object}}
            relations: {type: array, items: {type: object}}
            sources: {type: array, items: {$ref: '#/components/schemas/SourceAsset'}}
    SourceAsset:
      type: object
      required: [id, sourceType, url, title, accessedAt]
      properties:
        id: {type: string}
        sourceType: {type: string}
        url: {type: string}
        libraryId: {type: string}
        relativePath: {type: string}
        contentHash: {type: string, pattern: '^[0-9a-f]{64}$'}
        title: {type: string}
        accessedAt: {type: string, format: date}
        locator: {type: string}
    CatalogImport:
      type: object
      required: [importId, libraryId, proposalHash, status, proposal, validation, diff, createdAt]
      properties:
        importId: {type: string}
        libraryId: {type: string}
        proposalHash: {type: string}
        status: {type: string, enum: [previewed, invalid, conflict, applied, rejected]}
        proposal: {type: object}
        validation: {type: object}
        diff: {type: object}
        createdAt: {type: string, format: date-time}
        appliedAt: {type: [string, 'null'], format: date-time}
        rejectedAt: {type: [string, 'null'], format: date-time}
    UpdateTrainingSettings:
      type: object
      required: [questionCount, newBudget, backlogPauseThreshold, overduePauseDays, pendingCandidateTtlHours, reportingTimeZone]
      properties:
        questionCount: {type: integer, minimum: 1, maximum: 100}
        newBudget: {type: integer, minimum: 0, description: 必须小于或等于 questionCount}
        backlogPauseThreshold: {type: integer, minimum: 0, maximum: 10000}
        overduePauseDays: {type: integer, minimum: 1, maximum: 365}
        pendingCandidateTtlHours: {type: integer, minimum: 1, maximum: 8760}
        reportingTimeZone: {type: string, example: Asia/Shanghai}
    TrainingSettings:
      allOf:
        - $ref: '#/components/schemas/UpdateTrainingSettings'
        - type: object
          required: [reviewBudget, updatedAt]
          properties:
            reviewBudget: {type: integer, minimum: 0, description: questionCount - newBudget}
            updatedAt: {type: string, format: date-time}
            updatedBy: {type: [string, 'null']}
    GenerationProfile:
      type: object
      required: [questionType, difficulty, knowledgePoint]
      properties:
        questionType:
          type: string
          enum: [single_choice, multiple_choice]
        difficulty:
          type: integer
          minimum: 1
          maximum: 5
        knowledgePoint:
          type: object
          required: [topicId, name, importance, applicableVersions, keywords, sourceRefs, sourceReferences, referenceLibraryIds]
          properties:
            topicId: {type: string}
            name: {type: string}
            importance: {type: integer, minimum: 1, maximum: 5}
            applicableVersions: {type: array, items: {type: string}}
            javaVersions:
              type: array
              deprecated: true
              description: 向后兼容字段；新客户端使用 applicableVersions。
              items: {type: string}
            keywords: {type: array, items: {type: string}}
            sourceRefs: {type: array, items: {type: string}}
            sourceReferences:
              type: array
              items: {$ref: '#/components/schemas/SourceAsset'}
            referenceLibraryIds:
              type: array
              items: {type: string}
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1}